[progress] refactored megaMenu to include SubMenu component
